requirements.txt
streamlit
pyodbc
python-dotenv
bcrypt